fs: fix+expose count_open_files()
Previously, count_open_files() overestimated (via rounding up to the
nearest BITS_PER_LONG) the number of open files. In future patches we're
going to need the exact number, so fix this behavior.

The only caller of count_open_files() currently is sane_fdtable_size(),
which does its own alignment anyway, so this Should Be Fine.

We also expose it for use in the misc cg in the next patch.

Signed-off-by: Tycho Andersen <tandersen@netflix.com>