perf build: fix out of tree build
It seems that a previous modification to sysreg-defs, which corrected
emitting the headaer to the specified output directory, exposed missing
subdir, prefix variables. This breaks out of tree builds of perf as the
file is now built into the output directory, but still tries to descend
into output directory as a subdir.

Fixes: a29ee6aea7
Signed-off-by: Ethan Adams <eadams@netflix.com>